+++ src/models/modelArtifact.ts
