+++ src/routes/sessionRoute.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { newSession } from "../controllers/session.js";
+import { newSession } from "../controllers/sessionController.js";
 import { verifyData } from "../middlewares/verifyData.js"
 import { authSchema } from "../schemas/auth.js";
